+++ aws_lambda_powertools/event_handler/openapi/constants.py
@@ -1,0 +1,2 @@
+DEFAULT_API_VERSION = "1.0.0"
+DEFAULT_OPENAPI_VERSION = "3.1.0"
